+++ data/src/main/java/gr/tsambala/tutorbilling/data/database/TutorBillingDatabase.kt
@@ -13,13 +13,14 @@ import gr.tsambala.tutorbilling.data.model.Student
 
 @Database(
     entities = [Student::class, Lesson::class],
-    version = 9, // Current version
+    version = 10, // Current version
     exportSchema = true,
     autoMigrations = [
         AutoMigration(from = 5, to = 6, spec = AutoMigration5To6::class),
         AutoMigration(from = 6, to = 7),
         AutoMigration(from = 7, to = 8),
-        AutoMigration(from = 8, to = 9)
+        AutoMigration(from = 8, to = 9),
+        AutoMigration(from = 9, to = 10)
     ]
 )
 abstract class TutorBillingDatabase : RoomDatabase() {